+++ components/Dashboard/styleslight.module.scss
@@ -80,3 +80,9 @@
     align-items: center;
     cursor: pointer;
 }
+
+@media only screen and (max-width: 700px) {
+    .tooltip .tooltiptext {
+        margin-left: -2rem;
+    }
+}
