+++ components/SearchBar/index.tsx
@@ -1,5 +1,5 @@
 import * as tyron from 'tyron'
-import React, { useState, useCallback, useEffect } from 'react'
+import React, { useState, useCallback } from 'react'
 import { useRouter } from 'next/router'
 import { toast } from 'react-toastify'
 import { useSelector } from 'react-redux'
@@ -20,14 +20,11 @@ import { updateResolvedInfo } from '../../src/store/resolvedInfo'
 import { updatePrev } from '../../src/store/router'
 import smartContract from '../../src/utils/smartContract'
 import toastTheme from '../../src/hooks/toastTheme'
-import useArConnect from '../../src/hooks/useArConnect'
 
 function Component() {
     const zcrypto = tyron.Util.default.Zcrypto()
     const Router = useRouter()
-    const { connect } = useArConnect()
     const net = useSelector((state: RootState) => state.modal.net)
-    const loginInfo = useSelector((state: RootState) => state.modal)
     const isLight = useSelector((state: RootState) => state.modal.isLight)
     const styles = isLight ? stylesLight : stylesDark
     const [name, setName] = useState('')
@@ -63,7 +60,6 @@ function Component() {
         key,
     }: React.KeyboardEvent<HTMLInputElement>) => {
         if (key === 'Enter') {
-            checkArConnect()
             if (name !== '') {
                 updatePrev(window.location.pathname)
                 getResults(name, domx)
@@ -390,12 +386,6 @@ function Component() {
             })
     }
 
-    const checkArConnect = () => {
-        if (loginInfo.arAddr) {
-            connect()
-        }
-    }
-
     return (
         <div className={styles.container}>
             <div className={styles.searchDiv}>
@@ -412,7 +402,6 @@ function Component() {
                     <div className={styles.bar} />
                     <div
                         onClick={() => {
-                            checkArConnect()
                             updatePrev(window.location.pathname)
                             if (name !== '') {
                                 getResults(name, domx)
